Move bulkhead gauge initialization to constructor
Since the Lazy used for bulkheads uses PublicationOnly, high-load access
on the bulkhead will cause several holders to get created. If the holder
initializes the gauge timer, the timers for the holders that race but
don't get set as the Lazy's value don't get immediately garbage collected,
and continue to fire.

Moving the gauge to the factory and having it loop over bulkheads is more
predictable.